feat(tui): blinking
