+++ tauri-runner/src/components/Sidebar/sidebar.css
@@ -1,0 +1,36 @@
+a {
+    text-transform: uppercase;
+    text-decoration: none;
+    color: #e7e8f2;
+    font-size: 13px;
+    opacity: 0.8;
+}
+a:hover {
+    opacity: 1;
+}
+
+li {
+    list-style-type: none;
+}
+
+.sidebar-items {
+    width: 203px;
+    padding: 0;
+    margin: 45px 0 0;
+    visibility: visible;
+    /* animation: linear 260ms hidden; */
+}
+
+.sidebar-item {
+    align-items: center;
+    display: flex;
+    box-sizing: content-box;
+    height: 32px;
+    padding: 0 16px 0 16px;
+    vertical-align: middle;
+    cursor: pointer;
+}
+
+.sidebar-item:hover {
+    background-color: rgba(44,45,82, .5);
+}
