+++ tauri-runner/src/pages/runner/runner.css
@@ -1,0 +1,57 @@
+.container {
+    margin: 0;
+    padding-top: 10vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    text-align: center;
+}
+
+
+input {
+    width: 60%;
+}
+
+input,
+button {
+    border-radius: 8px;
+    border: 1px solid transparent;
+    padding: 0.6em 1.2em;
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    color: #0f0f0f;
+    background-color: #ffffff;
+    transition: border-color 0.25s;
+    box-shadow: 0 2px 2px rgba(0, 0, 0, 0.2);
+}
+
+button {
+    cursor: pointer;
+}
+
+button:hover {
+    border-color: #396cd8;
+}
+
+input,
+button {
+    outline: none;
+}
+
+.textField {
+    width: 600px;
+}
+
+.alertPos {
+    width: 600px;
+    transform: translateY(5px);
+}
+
+.btnPos {
+    transform: translateY(10px);
+}
+
+.nextButtonsG {
+    padding-top: 10px;
+}
